apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
geoalchemy2==0.15.2
redis==5.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
loguru==0.7.2
python-multipart==0.0.12
websockets==13.1
